requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
requests
reportlab
Pillow
python-dotenv
gunicorn
